spot-scheduler/requirements.txt
requests==2.31.0
paho-mqtt==2.1.0 